Change: Pass fn that creates mock of user interface into global mocks
